+++ Changelog.md
@@ -1,4 +1,8 @@
 ## [Unreleased]
+### Update
+- Set args: end_entity_profile_name, cert_profile_name as None
+  for functions: generate_certificate_pkey_on_server, generate_certificate
+
 
 ## [1.0.2]
 ### Fix